Really adjust namespaces.
