+++ VSDiagnostics/VSDiagnostics/VSDiagnostics.Test/VSDiagnostics.Test.csproj
@@ -149,8 +149,8 @@
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Tests\Utilities\ExtensionsTests.cs" />
     <Compile Include="Tests\Utilities\NamingConventionsTests.cs" />
-    <Compile Include="Tests\XMLDocComments\RedundantXMLDocParameter\RedundantXMLDocParameterTests.cs" />
-    <Compile Include="Tests\XMLDocComments\RedundantXMLDocReturn\RedundantXMLDocReturnTests.cs" />
+    <Compile Include="Tests\XMLDocumentation\RedundantXMLDocParameter\RedundantXMLDocParameterTests.cs" />
+    <Compile Include="Tests\XMLDocumentation\RedundantXMLDocReturn\RedundantXMLDocReturnTests.cs" />
   </ItemGroup>
   <ItemGroup>
     <None Include="app.config">
