+++ VSDiagnostics/VSDiagnostics/VSDiagnostics/Diagnostics/Async/SyncMethodWithAsyncSuffix/SyncMethodWithAsyncSuffixAnalyzer.cs
@@ -1,12 +1,11 @@
 ﻿using System.Collections.Immutable;
-using System.Linq;
 using Microsoft.CodeAnalysis;
 using Microsoft.CodeAnalysis.CSharp;
 using Microsoft.CodeAnalysis.CSharp.Syntax;
 using Microsoft.CodeAnalysis.Diagnostics;
 using VSDiagnostics.Utilities;
 
-namespace VSDiagnostics.Diagnostics.Async.SyncMethodWithSyncSuffix
+namespace VSDiagnostics.Diagnostics.Async.SyncMethodWithAsyncSuffix
 {
     [DiagnosticAnalyzer(LanguageNames.CSharp)]
     public class SyncMethodWithAsyncSuffixAnalyzer : DiagnosticAnalyzer
